Moved data cache to a more localed utilities location
